revive sd_card, storage
