+++ omi/firmware/omi/CMakeLists.txt
@@ -20,6 +20,11 @@ file(GLOB dk2_sources
     src/lib/dk2/transport.c
     src/lib/dk2/button.c
 )
+
+if(CONFIG_OMI_ENABLE_OFFLINE_STORAGE)
+    list(APPEND dk2_sources src/lib/dk2/storage.c)
+endif()
+
 target_sources(app PRIVATE ${dk2_sources} ${app_sources})
 
 if(CONFIG_OMI_CODEC_OPUS)
