+++ omi/firmware/omi/omi.conf
@@ -306,7 +306,7 @@ CONFIG_FAT_FILESYSTEM_ELM=y
 CONFIG_FS_FATFS_MOUNT_MKFS=y
 CONFIG_FS_FATFS_EXFAT=y
 #nessessary?
-CONFIG_MAIN_STACK_SIZE=2048
+CONFIG_MAIN_STACK_SIZE=8192
 # CONFIG_DISK_DRIVER_SDMMC=y
 # CONFIG_SPI=y
 # CONFIG_HEAP_MEM_POOL_SIZE=512 # disabled by conflicting
@@ -361,7 +361,7 @@ CONFIG_PM_DEVICE=y
 # Omi features configuration
 #
 CONFIG_OMI_CODEC_OPUS=y
-CONFIG_OMI_ENABLE_OFFLINE_STORAGE=n
+CONFIG_OMI_ENABLE_OFFLINE_STORAGE=y
 CONFIG_OMI_ENABLE_ACCELEROMETER=n
 CONFIG_OMI_ENABLE_BUTTON=y
 CONFIG_OMI_ENABLE_SPEAKER=n
